Рефактор структуры компонентов эпика. TaskInlined, TaskGroupList, HarnessView, RightContainer, LeftContainer и TaskList разделены на отдельные компоненты. Убраны лишние комментарии и заглушки api.
