+++ src/components/epic/leftContainer/LeftContainer.jsx
@@ -1,51 +1,14 @@
 import React from "react";
-import {Typography} from "@material-ui/core";
-import {useLeftContainerStyles} from "./LeftContainerStyles";
-import clsx from "clsx";
-import {EpicInfo} from "./epicInfo/EpicInfo";
-import {SharedWith} from "./sharedWith/SharedWith";
-import {Clear, Create, Done} from "@material-ui/icons";
-import EdiText from 'react-editext'
+import {LeftContainerView} from "./LeftContainerView";
 
 export const LeftContainer = (props) => {
-    const classes = useLeftContainerStyles();
     const handlerRenameEpic = () => {
-
     };
     return (
-        <div className={clsx(classes.epicSideBar, classes.flexColumn)}>
-            <Typography
-            variant="h5"
-            align="center">
-            <EdiText
-                value={props.epicDetails.epicName === null ? "name is null" : props.epicDetails.epicName}
-                type="text"
-                onSave={handlerRenameEpic}
-                editButtonClassName={classes.renameButton}
-                editButtonContent={<Create/>}
-                saveButtonContent={<Done/>}
-                saveButtonClassName={classes.renameButton}
-                cancelButtonContent={<Clear/>}
-                cancelButtonClassName={classes.renameButton}
-                hideIcons
-                showButtonsOnHover
-                cancelOnUnfocus
-                cancelOnEscape
-                submitOnEnter
-            />
-            </Typography>
-            <Typography
-                variant="body1"
-            >
-                {props.epicDetails.epicDescription}
-            </Typography>
-            <EpicInfo
-                totalTasks={props.epicDetails.totalTasks}
-                totalTaskLists={props.epicDetails.totalTaskLists}
-            />
-            <SharedWith
-                style={classes.flexForceBottom}
-            />
-        </div>
+        <LeftContainerView
+            epicDetails={props.epicDetails}
+            epicId={props.epicId}
+            handlerRenameEpic={handlerRenameEpic}
+        />
     )
 }
