+++ src/components/epic/leftContainer/epicInfo/EpicInfo.jsx
@@ -3,12 +3,11 @@ import Card from "@material-ui/core/Card";
 
 export const EpicInfo = (props) => {
     return (
-        <Card
-        >
+        <Card>
             Epic info will go here<br/>
             Total Task Lists: {props.totalTaskLists}<br/>
             Total Tasks: {props.totalTasks}<br/>
-            etc.
+            Etc.
         </Card>
     )
 }
