+++ src/components/epic/rightContainer/TabCoomponents/GrantChartTab.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import {TabPanel} from "./TabPanel";
+import {GrantChart} from "../../../grantChart/GrantChart";
+
+export const GrantChartTab = (props) => {
+    return (
+        <TabPanel value={props.selectedLayout} index={props.index} classes={props.classes}>
+            <GrantChart/>
+        </TabPanel>
+    )
+}
