+++ src/components/taskInlined/taskComponents/TaskShared.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import {AccountCircle} from "@material-ui/icons";
+import {useTaskInlinedStyles} from "../TaskInlinedStyles";
+
+export const TaskShared = (props) => {
+    const classes = useTaskInlinedStyles();
+    return (
+        <AccountCircle className={classes.flexForceRight}/>
+    )
+}
