+++ src/components/taskInlined/taskComponents/TaskTimer.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import {PlayArrow} from "@material-ui/icons";
+
+export const TaskTimer = (props) => {
+    return (
+        <PlayArrow/>
+    )
+}
