+++ src/components/taskList/AddTaskForm.jsx
@@ -3,9 +3,6 @@ import {useFormik} from "formik";
 import {createTask} from "../../api/EpicApi";
 import OutlinedInput from "@material-ui/core/OutlinedInput";
 import {useTaskListStyles} from "./TaskListStyles";
-import clsx from "clsx";
-import InputLabel from "@material-ui/core/InputLabel";
-
 
 export const AddTaskForm = (props) => {
     const classes = useTaskListStyles();
@@ -19,7 +16,6 @@ export const AddTaskForm = (props) => {
         }
     });
     const handleAddTask = taskName => {
-        console.log(taskName);
         addTask(taskName, props.setTasks, props.epicId, props.taskListId, props.tasks);
     };
     return (
